ref(Makefile): if shell installed - install complele

+++ Makefile
@@ -12,12 +12,16 @@ all: install
 
 install:
 	install -Dm755 $(SCRIPTNAME) $(DESTDIR)$(PREFIX)/bin/$(BINARY)
-	install -vd "$(DESTDIR)$(BASHCOMPDIR)" && install -m 0644 note.bash-completion "$(DESTDIR)$(BASHCOMPDIR)/note"
-	install -vd "$(DESTDIR)$(ZSHCOMPDIR)" && install -m 0644 note.zsh-completion "$(DESTDIR)$(ZSHCOMPDIR)/_note"
 	install -vd "$(DESTDIR)$(MANDIR)/man1" && install -m 0644 manpage "$(DESTDIR)$(MANDIR)/man1/note.1"
+	@if which zsh &>/dev/null; then \
+        install -vd "$(DESTDIR)$(ZSHCOMPDIR)" && install -m 0644 note.zsh-completion "$(DESTDIR)$(ZSHCOMPDIR)/_note"; \
+    fi
+	@if which bash &>/dev/null; then \
+		install -vd "$(DESTDIR)$(BASHCOMPDIR)" && install -m 0644 note.bash-completion "$(DESTDIR)$(BASHCOMPDIR)/note"; \
+    fi
 
 uninstall:
 	rm "$(DESTDIR)$(PREFIX)/bin/$(BINARY)"
-	rm "$(DESTDIR)$(BASHCOMPDIR)/note"
-	rm "$(DESTDIR)$(ZSHCOMPDIR)/_note"
 	rm "$(DESTDIR)$(MANDIR)/man1/note.1"
+	test -e "$(DESTDIR)$(BASHCOMPDIR)/note" && rm "$(DESTDIR)$(BASHCOMPDIR)/note"
+	test -e "$(DESTDIR)$(ZSHCOMPDIR)/_note" && rm "$(DESTDIR)$(ZSHCOMPDIR)/_note"
